Run request in interp if a breakpoint is set
Summary: We force a request to be run in interpreter if a breakpoint is set. Note that if a breakpoint is set while a request is running, we will not enter a TC after a BP is set, but the request may already be running in a TC, so the running request may not hot the breakpoint, but all subsequent requests are guaranteed to hit the breakpoint. Also, a breakpoint may not be resolved until a unit is loaded (it may not have the full path), so we need to track unresolved breakpoints as well.

Reviewed By: jeffreytan81

Differential Revision: D30533075

fbshipit-source-id: 2b02137e67526eb5d9e1c2cae06de6f71a85099c